Implemented BasicComponent in the bundling environment
Ported the component test case from runtime to a component suite and
then set it up to use the `BundlingRenderDelegate`.

To make the tests pass, the bundling tests implement a reference implementation
of the compile-time and runtime environments designed to work together. These
tests do *not* use the TestEnvironment, which is highly coupled to lazy mode.

The next steps are to finish porting over the rest of the suites that already
subclass AbstractRenderTest, and once done, port the rest of the component
tests to work with that system as suites.